requirements.txt
appdirs==1.4.3
appnope==0.1.0
arelle==2.2
attrs==19.3.0
backcall==0.2.0
beautifulsoup4==4.9.0
bleach==3.1.5
certifi==2020.4.5.2
chardet==3.0.4
colorama==0.4.3
cursor==1.3.4
cycler==0.10.0
dart-fss==0.3.0
DateTime==4.3
decorator==4.4.2
defusedxml==0.6.0
entrypoints==0.3
et-xmlfile==1.0.1
fake-useragent==0.1.11
halo==0.0.29
html5lib==1.0.1
idna==2.9
importlib-metadata==1.6.1
ipython==7.15.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isodate==0.6.0
jdcal==1.4.1
jedi==0.17.0
Jinja2==2.11.2
jsonschema==3.2.0
kiwisolver==1.2.0
log-symbols==0.0.14
lxml==4.5.0
MarkupSafe==1.1.1
matplotlib==3.2.1
mistune==0.8.4
mpl-finance==0.10.1
multitasking==0.0.9
nbconvert==5.6.1
nbformat==5.0.7
notebook==6.0.3
numpy==1.18.2
openpyxl==3.0.3
packaging==20.4
pandas==1.0.3
pandas-datareader==0.8.1
pandocfilters==1.4.2
parso==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
prometheus-client==0.8.0
prompt-toolkit==3.0.5
protobuf==3.12.2
ptyprocess==0.6.0
Pygments==2.6.1
pykrx==0.1.21
pymongo==3.10.1
pyparsing==2.4.7
pyrsistent==0.16.0
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.3.1
pyzmq==19.0.0
qtconsole==4.7.4
QtPy==1.9.0
requests==2.23.0
selenium==3.141.0
Send2Trash==1.5.0
six==1.15.0
soupsieve==2.0
spinners==0.0.24
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
tornado==6.0.4
tqdm==4.45.0
traitlets==4.3.3
urllib3==1.25.8
wcwidth==0.2.4
webencodings==0.5.1
wget==3.2
widgetsnbextension==3.5.1
xlrd==1.2.0
XlsxWriter==1.2.8
xmltodict==0.12.0
yfinance==0.1.54
zipp==3.1.0
zope.interface==5.1.0
